Spell the integer widths out and move ModuleConnector into the framework
unsigned, unsigned long long and long long are gone in favour of uint32_t, uint64_t
and int64_t, so a frame id and a millisecond count say how wide they are.

Seconds, Milliseconds and Microseconds live in TimeExtensions next to each other
now, and Stopwatch takes them from there rather than declaring its own copies. The
queue's bounded wait asks for Microseconds instead of reaching for std::chrono.

ModuleConnector knows about ports and modules and nothing about faces, so it sits in
Framework/Graph with the rest of them; it no longer pulls in OpenCV for an assert,
and ModuleGraph takes its predecessor map from there instead of declaring a second
one of its own.

+++ FaceApi/FaceApi.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <atomic>
+#include <cstdint>
 #include <memory>
 #include <string>
 
@@ -42,12 +43,12 @@ namespace face
 
     void OnOffVerbose();
 
-    inline unsigned GetLastFrameId() const
+    inline uint32_t GetLastFrameId() const
     {
       return mModuleGraph ? mModuleGraph->GetLastFrameId() : 0U;
     }
 
-    inline long long GetLastTimestamp() const
+    inline int64_t GetLastTimestamp() const
     {
       return mModuleGraph ? mModuleGraph->GetLastTimestamp() : 0LL;
     }
@@ -73,7 +74,7 @@ namespace face
 
     std::shared_ptr<ModuleGraph> mModuleGraph = nullptr;
 
-    std::atomic<unsigned> mCameraFrameId{ 0U };
+    std::atomic<uint32_t> mCameraFrameId{ 0U };
 
     MessageQueue mOutputQueue;
   };
